+++ packages/catalog/src/core/hooks/index.ts
@@ -1,3 +1,1 @@
 export * from './use-catalog';
-export * from './use-on-click-outside';
-export * from './use-window-event-listener';