+++ packages/catalog/src/core/hooks/use-on-click-outside/index.ts
@@ -1,1 +1,0 @@
-export { default as useOnClickOutside } from './use-on-click-outside';