+++ packages/navigation-bar/src/index.ts
@@ -1,0 +1,5 @@
+import NavigationBar from './navigation-bar';
+import NavigationBarLink from './navigation-bar-link';
+
+export * from './types';
+export { NavigationBar, NavigationBarLink };