+++ tooling/template-base/tailwind.config.js
@@ -36,6 +36,24 @@ module.exports = {
       '2xl': '1920px',
       '3xl': '2560px'
     },
+    boxShadow: {
+      DEFAULT: '0px 0px 5px #4d90fe',
+      lg: '0px 0px 12px -2px rgba(0, 0, 0, 0.25)',
+      none: 'none'
+    },
+    fontSize: {
+      xs: '0.625rem',
+      sm: '.875rem',
+      base: '1rem',
+      lg: '1.125rem',
+      xl: '1.25rem',
+      '2xl': '1.5rem',
+      '3xl': '1.875rem',
+      '4xl': '2.25rem',
+      '5xl': '3rem',
+      '6xl': '4rem',
+      '7xl': '5rem'
+    },
     extend: {
       colors: {
         accent: getInstanceSetting('accentColor', '#111111'),