+++ packages/catalog/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.3](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/catalog@1.1.2...@thoughtindustries/catalog@1.1.3) (2022-05-04)
+
+**Note:** Version bump only for package @thoughtindustries/catalog
+
+
+
+
+
 ## [1.1.2](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/catalog@1.1.1...@thoughtindustries/catalog@1.1.2) (2022-04-08)
 
 **Note:** Version bump only for package @thoughtindustries/catalog