+++ packages/featured-content/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.2.0-alpha.2](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/featured-content@1.2.0-alpha.1...@thoughtindustries/featured-content@1.2.0-alpha.2) (2022-05-04)
+
+**Note:** Version bump only for package @thoughtindustries/featured-content
+
+
+
+
+
 # [1.2.0-alpha.1](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/featured-content@1.2.0-alpha.0...@thoughtindustries/featured-content@1.2.0-alpha.1) (2022-04-01)