+++ packages/testimonial/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.1](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/testimonial@1.1.0...@thoughtindustries/testimonial@1.1.1) (2022-05-04)
+
+**Note:** Version bump only for package @thoughtindustries/testimonial
+
+
+
+
+
 # 1.1.0 (2022-03-28)